+++ src/BazarServiceProvider.php
@@ -45,6 +45,7 @@ class BazarServiceProvider extends ServiceProvider
         Contracts\Gateway\Manager::class => Gateway\Manager::class,
         Contracts\Shipping\Manager::class => Shipping\Manager::class,
         Contracts\Repositories\TaxRepository::class => Repositories\TaxRepository::class,
+        Contracts\Repositories\AssetRepository::class => Repositories\AssetRepository::class,
         Contracts\Repositories\DiscountRepository::class => Repositories\DiscountRepository::class,
         Contracts\Repositories\ConversionRepository::class => Repositories\ConversionRepository::class,
     ];
@@ -165,6 +166,7 @@ class BazarServiceProvider extends ServiceProvider
             $this->commands([
                 Console\Commands\InstallCommand::class,
                 Console\Commands\PublishCommand::class,
+                Console\Commands\LinkAssetsCommand::class,
                 Console\Commands\ClearCartsCommand::class,
                 Console\Commands\ClearChunksCommand::class,
             ]);