+++ src/Http/Requests/OrderStoreRequest.php
@@ -24,8 +24,9 @@ class OrderStoreRequest extends FormRequest
             ],
             'products' => ['required', 'array'],
             'products.*.id' => ['required'],
-            'products.*.item_tax' => ['nullable', 'numeric', 'min:0'],
-            'products.*.item_quantity' => ['nullable', 'numeric', 'gt:0'],
+            'products.*.item.tax' => ['required', 'numeric', 'min:0'],
+            'products.*.item.price' => ['required', 'numeric', 'min:0'],
+            'products.*.item.quantity' => ['required', 'numeric', 'gt:0'],
             'discount' => ['nullable', 'numeric', 'min:0'],
             'currency' => [
                 'required',