+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.4.6 (2021-02-03)
+### Added
+- Added `AssetRepository` and `Asset` facade
+- Added `LinkAssetsCommand` command
+
+### Changed
+- Changed product addtion when creating orders manually
+
 ## v0.4.5 (2021-02-01)
 ### Fixed
 - Fixed Order creation error [#100](https://github.com/conedevelopment/bazar/issues/100)